requirements.txt
fpdf2==2.8.5
qrplatba==1.1.1